Prior to making javascript changes for win txt
The kinging method works needing the image to end with King specifically, all solid colors are labeled with numbers, Sam's designs are now (letter, letter) based. DJ's are being considered how to number. A working but not completely final leaderboard is added along with its CSS. A Separate folder is made to move the zip files that contain both the king pieces and the non-king pieces.